Exclude XP Gain from Blocked Channels
Prevent XP gain from blocked channels

Modified the MessageCreate event to exclude messages from blocked channels when calculating XP gain. If a message is from a bot or originates from a channel in the blocked list, XP is not awarded. Additionally, added methods to the Guild class for managing blocked channels, including adding, removing, and retrieving the list of blocked channels.